+++ test/workhorse/performer_test.rb
@@ -13,4 +13,20 @@ class Workhorse::WorkerTest < WorkhorseTest
     assert_equal 2, DbConnectionTestJob.db_connections.count
     assert_equal 2, DbConnectionTestJob.db_connections.uniq.count
   end
+
+  def test_on_exception
+    old_callback = Workhorse.on_exception
+    exception = nil
+
+    Workhorse.on_exception = proc do |e|
+      exception = e
+    end
+
+    Workhorse.enqueue FailingTestJob.new
+    work 0.2, polling_interval: 0.2
+
+    assert_equal exception.message, FailingTestJob::MESSAGE
+  ensure
+    Workhorse.on_exception = old_callback
+  end
 end